event_management_microservice_mysql/backend/requirement.txt
configparser==5.3.0
requests==2.31.0
urllib3==2.0.3
Flask==2.3.2
virtualenv==20.23.1
waitress==2.1.2
mysql-connector-python==8.1.0
